GPS code for arduino
GPS arduino code with loop() and setup() functions that initializes the SD card and puts data formatted into a single csv file. 

The gps processes it through a software serial port using UART communication protocol and uses a UBX library to simplify the code syntax and print to the csv file.