Integrate OpEn solver for MPC optimization
Added the optimization_engine crate and implemented an OpEn solver for the MPC problem, replacing the previous SQP/PGD approach in main.rs. Updated reference setpoints, cost weights, and plotting colors.

Issues:
OpEn is slower, most likely due to calculation of Jacobian for the cost gradient. Will need to look into matrix derivatives of the quadratic program for the gradient.