Refactor MPC QP solver and tune optimization parameters
Replaces the old trajectory cost and gradient with new QP cost and gradient functions. Replace automatic differentiation via finite difference with the analytical derivative of the quadratic program's objective function. Tune solver parameters based on max number of iterations and tolerance.

Issues:
Is faster than the old implementation for solving but the control input for thrust is noisier. May consider changing max number of iterations and tolerance to get better control but this will likely result in a slow down in computation.